chore: Update Flask application to listen on specified port
This commit updates the Flask application to listen on a specified port. Previously, the application was running on the default port 5000. Now, the port is determined by the `PORT` environment variable, with a fallback to port 5000 if the variable is not set. This change allows for more flexibility in deploying the application to different environments.